requirements.txt
flask
flask-cors
torch
transformers
datasets
peft
accelerate
sentencepiece
protobuf
python-dotenv
google-auth
requests
PyJWT
cryptography
google-generativeai
httpx
PyPDF2
langdetect
APScheduler
SQLAlchemy
Pillow
